requirements.txt
aiohttp>=3.9.0
aiosqlite>=0.19.0
lxml>=5.0.0
beautifulsoup4>=4.12.0
pyyaml>=6.0.0
apscheduler>=3.10.0
tenacity>=8.2.0
fake-useragent>=1.4.0
python-dotenv>=1.0.0
supabase>=2.0.0
httpx>=0.25.0
robotexclusionrulesparser>=1.7.1
python-dateutil>=2.8.0
fastapi>=0.109.0
uvicorn>=0.27.0
